adding additional helpers and cleaning up some tests
